renamed FilesTools', Parser's and GraphVertex' ints

+++ src/baseStructures/graph/GraphVertex.hpp
@@ -139,7 +139,7 @@ public:
   /// OrientedGraph methods
   /// @param i_indent
   /// @return
-  std::string toGraphML(int i_indent = 0, std::string i_prefix = "") const;
+  std::string toGraphML(uint16_t i_indent = 0, std::string i_prefix = "") const;
 
   /// @brief This method is used as a substructure for
   /// OrientedGraph methods
